requirements.txt
anthropic>=0.40.0
openai>=1.0.0
pdfplumber>=0.11.0
httpx>=0.27.0
networkx>=3.3
python-louvain>=0.16
pyvis>=0.3.2
pydantic>=2.0
python-dotenv>=1.0
rich>=13.0
sqlite-utils>=3.36
pyyaml>=6.0
pytest>=8.0
pytest-asyncio>=0.23
